+++ Environments.EnvironmentMappings/EnvironmentProfileDefaults.cs
@@ -7,10 +7,11 @@ public static class EnvironmentProfileDefaults
         {
             new(EnvironmentProfileNames.Development, CanonicalEnvironment.Development),
             new(EnvironmentProfileNames.Local, CanonicalEnvironment.Development),
+            new(EnvironmentProfileNames.Cde, CanonicalEnvironment.Development),
             new(EnvironmentProfileNames.Staging, CanonicalEnvironment.Staging),
             new(EnvironmentProfileNames.Qa, CanonicalEnvironment.Staging),
+            new(EnvironmentProfileNames.QualityAssurance, CanonicalEnvironment.Staging),
             new(EnvironmentProfileNames.Uat, CanonicalEnvironment.Staging),
-            new(EnvironmentProfileNames.Cde, CanonicalEnvironment.Staging),
             new(EnvironmentProfileNames.PreProd, CanonicalEnvironment.Staging),
             new(EnvironmentProfileNames.Production, CanonicalEnvironment.Production)
         }.AsReadOnly();