+++ Environments.EnvironmentMappings/EnvironmentProfileNames.cs
@@ -7,6 +7,7 @@ public static class EnvironmentProfileNames
     public const string Production = "Production";
     public const string Local = "Local";
     public const string Qa = "QA";
+    public const string QualityAssurance = "QualityAssurance";
     public const string Uat = "UAT";
     public const string Cde = "CDE";
     public const string PreProd = "PreProd";